requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.3
django-celery-beat==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jira==3.4.1
kombu==5.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
nose==1.3.7
ntlm-auth==1.5.0
oauthlib==3.2.2
packaging==23.0
prompt-toolkit==3.0.36
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
svn==1.0.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
